[feat]admin feedback toggle

+++ src/main/java/com/gg/server/admin/feedback/data/FeedbackAdminRepository.java
@@ -3,5 +3,5 @@ package com.gg.server.admin.feedback.data;
 import com.gg.server.domain.feedback.data.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackAdminRepository extends JpaRepository<Feedback, Integer> {
+public interface FeedbackAdminRepository extends JpaRepository<Feedback, Long> {
 }
